Add SP800155 event generation.
This produces another file alongside the snapshotted firmware and
signature for the VMM to provide as a boot-services UEFI variable.
The firmware will read this and extend the HOB list in the PEI stage to
make Tcg2Dxe emit the EV_NO_ACTION events for verifiers to read and find
where the reference values can be found.

Changes to existing marshaling logic comes from debugging an end-to-end
test.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>