Fix ByteSizeArray to CStr and drop uint16 size
Upstream doesn't like the UINT16 event size prefix to make ingestion
easier, so now the ~700 line parser is in review for EDK2.

The Byte strings are all expected to be 0-terminated C strings, so I've
added that to the marshaling and unmarshaling logic. It's unclear if a
size 0 string is allowed since the string is expected to be
null-terminated, so I've gone with expecting to emit a null for empty
string.