Fix measurement calculation following Vanadium fix
Vanadium did not respect the TDVF Metadata attribute field for whether a
region should be measured or not. By properly accounting for the
extendmr bit in TDVF metadata sections in the hypervisor, the RAM banks
data in the TDHOB section are not relevant to MRTD.

The CFV and TDHOB will still be relevant to RTMRs.