Account for TDX PermMem and Payload section types
The TDVF specification defines 8 section types, and we only covered 4
before. The Payload, PayloadParam, and PermMem sections are all
important pieces to account for running a measured Shim or container.
The TD_INFO section is not yet relevant, but I've added the constant
nonetheless. It is a pointer into the BFV itself, so it does not account
for any measurement.

Avoid mutating append and comment tdhob verification logic.